Seeds stay within what the runtimes take
--seed took any u64, but sd-cli reads a seed as a signed 64-bit number and
stops without a word beyond it, which came out as an environment problem
that sent people to doctor. And each picture's seed was the first plus its
place, which overflowed: a bug report in a debug build, a silent wrap to 0 in
release. `image crop --box` overflowed the same way in its own error message.

--seed (gen and theme) now stops at 9223372036854775807 with a sentence that
says so; a batch brief's seed is checked when the briefs are read; a run of
seeds that would pass the largest is refused; and the engine checks the seed
before it starts the runtime.

+++ crates/folderskin-cli/src/images.rs
@@ -212,8 +212,8 @@ fn crop(args: &CropArgs, out: &Arc<Out>) -> Result<(), CliError> {
                 "That box goes outside the picture.",
                 format!(
                     "The picture is {w} × {h}; the box ends at {} × {}.",
-                    x + bw,
-                    y + bh
+                    u64::from(x) + u64::from(bw),
+                    u64::from(y) + u64::from(bh)
                 ),
             )
             .fix("Keep X + WIDTH and Y + HEIGHT within the picture."));
